refactor file_transfer dir
